rename traceback script to avoid conflicts

+++ webapp/lib/needleman_wunsch.py
@@ -1,5 +1,5 @@
 #from traceback import _get_paths, _modify_path
-from webapp.lib.traceback import _get_paths, _modify_path
+from webapp.lib.pairwise_traceback import _get_paths, _modify_path
 
 # Gibt die Richtung/Vorgänger zurück
 def traceback_value(dia,ver,hor, similarity : bool):
@@ -61,7 +61,7 @@ def get_alignment(seq1 : str, seq2: str, path : list):
 
 
 # Needlemann-Wunsch-Algorithmus
-def get_result(sequence1: str, sequence2: str, match: float, mismatch: float, gap_penalty: float, similarity : bool):
+def get_result(sequence1: str, sequence2: str, match : int, mismatch : int, gap_penalty : int, similarity : bool):
     seq1,seq2 = '-' + sequence1, '-' + sequence2
 
     traceback = [[None for j in range(len(seq2))] for i in range(len(seq1))]
@@ -116,6 +116,7 @@ def get_result(sequence1: str, sequence2: str, match: float, mismatch: float, ga
     print('alignments')
     for i in alignments: print(i)
     '''
+    
     return {'matrix': matrix,
             'alignments': alignments,
             'score': score,
